refactor: ValidityMessage 를 컴포넌트 항목으로 이동
